requirements.txt
pyqrcode
python-telegram-bot
pypng